Карточка темы выделена в отдельный компонент
