+++ packages/client/src/pages/forum/components/index.ts
@@ -1,0 +1,1 @@
+export { default as ThemeCard } from './ThemeCard/ThemeCard'
